To BigInt values changed

+++ EpamLabTasksPart2/TaskWithPointAndLinq/IntExtension.cs
@@ -3,16 +3,17 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Numerics;
 
 namespace TaskWithPointAndLinq
 {
   public static  class IntExtension
     {
-        public static bool IsIntSimple(this ulong x)
+        public static bool IsIntSimple(this BigInteger x)
         {
-            for (int i = 2; i <= (int)(x / 2); i++)
+            for (int i = 2; i <= (x / 2); i++)
             {
-                if ((int)x % i == 0)
+                if (x % i == 0)
                     return false;
             }
             return true;
